Feat: Adding and improving menu
Adding menu animations for hovering over each item and improving the menu/cross button animation for consistency. Also adding a click off area so the rest of the site cannot be accessed at the same time as the menu and the menu is easy to close if the user does not notice the X button. TODO: fix unerline animations to they only take up the width of the text and not the width of the menu-items div, improve click off region by having it not cover the menu and fade in and out for smooth transition