Refactor/Feat: replacing cube effect with regular scroll
Done so that all the information is on a single page and people don't need to navigate to a bunch of different pages when on the site. Also made menu items scroll to the locations of the relevant slides instead of loading a new page. Each slide looks kinda scuffed right now, need to dedicate time to formatting each slide and adding all the necessary information to each one too